+++ branch/settings/base.py
@@ -32,6 +32,7 @@ import sys
 from datetime import datetime
 
 import environ
+from django.utils.translation import gettext_lazy as _
 
 # Build paths inside the project like this: BASE_DIR('subdir').
 BASE_DIR = environ.Path(__file__) - 3
@@ -90,6 +91,8 @@ IGNORABLE_404_URLS = (
 
 # Application definition
 
+SITE_NAME = _('Serghei Iakovlev')
+
 # The database ID of the django.contrib.sites.models.Site
 # object associated with that particular settings file.
 SITE_ID = 1
@@ -117,6 +120,7 @@ THIRD_PARTY_APPS = (
 # Apps specific for this project go here.
 LOCAL_APPS = (
     'apps.core',
+    'apps.seo',
     'apps.website',
     'apps.trumbowyg',
     'apps.blog',
@@ -169,8 +173,9 @@ TEMPLATES = [
                 'apps.core.context_processors.i18n',
                 'apps.core.context_processors.global_settings',
 
+                'apps.seo.context_processors.google_analytics',
+
                 'apps.website.context_processors.app_settings',
-                'apps.website.context_processors.base_url',
 
                 'apps.pages.context_processors.pages',
             ],
@@ -339,6 +344,7 @@ STATIC_ROOT = BASE_DIR('static')
 STATICFILES_DIRS = (
     BASE_DIR('assets'),
     NODE_MODULES_PATH,
+    BASE_DIR('apps', 'seo', 'assets'),
     BASE_DIR('apps', 'trumbowyg', 'assets'),
     BASE_DIR('apps', 'telegraph', 'assets'),
 )
@@ -381,3 +387,7 @@ EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'
 
 FEED_MAX_ITEMS = 20
 FEED_WORD_LIMIT = 30
+
+# SEO Tools
+
+GA_TRACKING_ID = env.str('GA_TRACKING_ID', None)